refactor(materializer): extract operation-index foundation (file #3)
Lift the OperationIndex/OperationIndexPatchResult types + buildOperationIndex +
findOperationForTranscriptSourceEntry into operation-index.ts (shared foundation
threaded through ~28 sites). Net 74/74, check 0.
